Android:Remove TableLayout, Fxed Padding, Small Clean ups [fixed]
Includes cleanup of all the pointless whitespace changes and leftovers
from trying to implement a copy-paste adapter from a tutorial.

Android:Fix:Fixed Padding

Android: Seperate Activity For FilePickers/External Emulator Launch

Removing the LinearLayout

This was originally labeled “redundant” but it is only unused because
the code written before this, but not submitted until after, did not
account for intermediate changes. The label, in itself, is redundant.

Android:Set Column Width Needed For autocolumn to work