FoldableLayout, Remove unused variables, Annotate hacks
This adds the new folding view for game details.
ExpandableHeightGridView was labeled a hack for an issue that had
“never been seen” and was later resolved by a this alternate hack.
Setting a variable to use as a parameter, and the parameter itself, are
unnecessary when the adapter has access to the original method.
All instances of a gameID should be updated at once, but the serial
should only be associated when either matched or null in the db.