Add proprietary Android TV layout
A width (equivalent to a single column) is required for the menu to
avoid recalculating the height and width after the view is populated.

The navigation background has been rerouted as the menu background to
differentiate it from the GridView and still maintain a unified pattern.

Restore toolbar support to the Android TV layout
Fix a whitespace tragedy. Thanks Xcode!
Clean up the code to reuse blocks as methods